Create README.md
In this code, we use the Math.random() function, which randomly returns a number between 0 and 1. Then we multiply it by 10 and round the result using the Math.floor() function to get a random integer between 0 and 9. Finally , we got 1 to the result to get a number between 1 and 10.

The alert() function is used to display the value of randomNumber in the browser window as a display message. You can add this code to an HTML document inside its <script> tag, or include it in the console. Upon launch, you found a random number in your browser window.